GEENOME

TGEENOME is a 16S rRNA metagenomic analysis tool that encompasses nearly all steps of downstream analysis. These steps include preprocessing to filter zero-inflated and low abundance data, exploring the dataset through various plots, calculating alpha and beta diversity, generating ordination plots, testing null hypotheses, and applying a novel differential expression method using generalized estimating equations. The results of all the aforementioned steps are compiled into a single, detailed PDF file within a well-organized folder structure.

GEENOME(
  physeq,
  variables,
  id,
  sample_var,
  group_var,
  out_cut,
  zero_cut,
  lib_cut,
  neg_lb,
  model_form,
  alpha,
  n_cl,
  prv_cut,
  AlphaBetaDiversity,
  color_label,
  BetaDiversity.distance,
  shape_label,
  axes,
  permanova.distance,
  permanova.strata,
  permanova.permutation_number,
  adj_pvalue,
  GEE_analysis,
  fill_Alpha
)

Arguments

physeq

Phyloseq format. The data input in phyloseq format.

variables

Character. The co-founders exist in metadata that affect on the taxa. in variables & cofounders for RCM

id

Character. Individual id (personal id for each participant even if repeated measures in time series)

sample_var

Character. The name of column storing sample IDs.

group_var

Character. The name of the group indicator.

out_cut

Numerical. fraction between 0 and 1.

zero_cut

Numerical. fraction between 0 and 1.

lib_cut

Numeric. Samples with library size less than lib_cut are not included in the analysis.

neg_lb

Logical. TRUE indicates a taxon would be classified as a structural zero.

model_form

Character. Just detect the variables you need to enter in the formula.

alpha

Numeric. alpha cutoff

n_cl

Numeric.

prv_cut

Numeric.

AlphaBetaDiversity

Logical. If you need to apply alpha and beta diversity steps or not. (the phyloseq must have tax_table)

color_label

Character. the variable or co-founder that used in coloring the plots. in the interested variable & in RCM

BetaDiversity.distance

Character. choose from "distanceMethodList" function

shape_label

Character. the variable or co-founder that used in shape of the plots. in RCM only

axes

Numeric. You need to detect the axes before plotting the ordination after scree plot

permanova.distance

Character. choose from "distanceMethodList" function

permanova.strata

Character. add strata if you need. The default is NULL

permanova.permutation_number

Numeric. The number of permutation.

adj_pvalue

Character. The choosen method in adjust p-value. choose from c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none")

GEE_analysis

Logical. Apply GEE downstream optionality if you don't need to re-analysis your data.

fill_Alpha

Character. choose your interested taxonomic level (Please make sure tax_table() in physeq has that's level)

Value

A PDF file in the same directory called "GEENOME_plots.pdf"

Details

GEENOME

Examples

library(phyloseq)

# load A two-week diet swap study between western (USA) and traditional (rural Africa) diets (Lahti et al. 2014).
data(dietswap, package = "microbiome")

phyloseq_data <- dietswap

# call library GEENOME
library(GEENOME)

# detect various parameters
# enter your phyloseq_data variable to be named "physeq"
physeq = phyloseq_data 

# make sure from the column names in sample_data(physeq)
variables <- c("bmi_group","timepoint.within.group") # variables & cofounders for RCM
color_label <- c("timepoint") # is the interested variable & in RCM
shape_label <- c("sex") # in RCM only
id = "subject" # individual id (personal id for each participant even if repeated measures in time series)
sample_var = "sample" # sample ID (equal to rownames of physeq@sam_data)

# Apply GEE downstream optionality if you don't need to re-analysis your data. 
GEE_analysis <- TRUE 
model_form <- c("bmi_group","timepoint.within.group")

# your need to make sure that phyloseq contain tax_table() before applying alpha & beta diversity
AlphaBetaDiversity = TRUE
BetaDiversity.distance <- "bray" # choose from these distances => distanceMethodList https://joey711.github.io/phyloseq/distance.html
fill_Alpha = "Family" # choose your interested taxonomic level (Please make sure tax_table() in physeq has that's level) 
axes = c(1,2) # You need to detect the axes before plotting the ordination after scree plot
permanova.distance <- c("bray") # https://rdrr.io/bioc/phyloseq/man/distanceMethodList.html
Permanova.strata = NULL # add strata if you need
permanova.permutation_number <- 9999

# Preprocessing parameters
group_var = NULL
out_cut = 0.05
zero_cut = 0.9
lib_cut = 1000
neg_lb = FALSE

adj_pvalue <- "BH" # choose from these => c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none")

# Run GEENOME
res <- GEENOME(physeq, variables, id, sample_var, group_var, out_cut, zero_cut, 
              lib_cut, neg_lb, model_form, alpha, n_cl, prv_cut,AlphaBetaDiversity,
              color_label,BetaDiversity.distance,shape_label,axes,permanova.distance,
              permanova.strata,permanova.permutation_number,adj_pvalue,GEE_analysis,fill_Alpha)
#> [1] "dplyr"
#> [1] ‘1.0.10’
#> [1] "nlme"
#> [1] ‘3.1.159’
#> [1] "ggplot2"
#> [1] ‘3.4.3’
#> [1] "compositions"
#> [1] ‘2.0.4’
#> Welcome to compositions, a package for compositional data analysis.
#> Find an intro with "? compositions"
#> [1] "plyr"
#> [1] ‘1.8.8’
#> ------------------------------------------------------------------------------
#> You have loaded plyr after dplyr - this is likely to cause problems.
#> If you need functions from both plyr and dplyr, please load plyr first, then dplyr:
#> library(plyr); library(dplyr)
#> ------------------------------------------------------------------------------
#> [1] "tidyverse"
#> [1] ‘1.3.2’
#> ── Attaching packages ─────────────────────────────────────── tidyverse 1.3.2 ──
#> ✔ tibble  3.1.8     ✔ purrr   0.3.5
#> ✔ tidyr   1.2.1     ✔ stringr 1.4.1
#> ✔ readr   2.1.3     ✔ forcats 0.5.2
#> ── Conflicts ────────────────────────────────────────── tidyverse_conflicts() ──
#> ✖ plyr::arrange()   masks dplyr::arrange()
#> ✖ nlme::collapse()  masks dplyr::collapse()
#> ✖ purrr::compact()  masks plyr::compact()
#> ✖ plyr::count()     masks dplyr::count()
#> ✖ plyr::failwith()  masks dplyr::failwith()
#> ✖ dplyr::filter()   masks stats::filter()
#> ✖ plyr::id()        masks dplyr::id()
#> ✖ dplyr::lag()      masks stats::lag()
#> ✖ plyr::mutate()    masks dplyr::mutate()
#> ✖ plyr::rename()    masks dplyr::rename()
#> ✖ plyr::summarise() masks dplyr::summarise()
#> ✖ plyr::summarize() masks dplyr::summarize()
#> [1] "gsubfn"
#> [1] ‘0.7’
#> [1] "zCompositions"
#> [1] ‘1.4.0.1’
#> 
#> Attaching package: ‘MASS’
#> The following object is masked from ‘package:dplyr’:
#> 
#>     select
#> 
#> Attaching package: ‘NADA’
#> The following object is masked from ‘package:compositions’:
#> 
#>     cor
#> The following object is masked from ‘package:stats’:
#> 
#>     cor
#> [1] "compositions"
#> [1] ‘2.0.4’
#> [1] "grid"
#> [1] ‘4.2.1’
#> [1] "gridExtra"
#> [1] ‘2.3’
#> [1] "nlme"
#> [1] ‘3.1.159’
#> [1] "optiscale"
#> [1] ‘1.2.2’
#> [1] "propr"
#> [1] ‘4.2.6’
#> [1] "webshot"
#> [1] ‘0.5.4’
#> [1] "ftExtra"
#> [1] ‘0.4.0’
#> [1] "flextable"
#> [1] ‘0.8.3’
#> [1] "caret"
#> [1] ‘6.0.93’
#> [1] "stringr"
#> [1] ‘1.4.1’
#> [1] "DT"
#> [1] ‘0.26’
#> [1] "htmlwidgets"
#> [1] ‘1.5.4’
#> [1] "geepack"
#> [1] ‘1.3.9’
#> [1] "ggpubr"
#> [1] ‘0.5.0’
#> [1] "vegan"
#> [1] ‘2.6.4’
#> This is vegan 2.6-4
#> [1] "scales"
#> [1] ‘1.2.1’
#> [1] "phyloseq"
#> [1] ‘1.40.0’
#> [1] "RCM"
#> [1] ‘1.12.0’
#> Registered S3 method overwritten by 'quantmod':
#>   method            from
#>   as.zoo.data.frame zoo 
#> [1] "data.table"
#> [1] ‘1.14.4’
#> [1] "microbiome"
#> [1] ‘1.18.0’
#> 
#> microbiome R package (microbiome.github.com)
#>     
#> 
#> 
#>  Copyright (C) 2011-2022 Leo Lahti, 
#>     Sudarshan Shetty et al. <microbiome.github.io>
#> [1] "heatmaply"
#> [1] ‘1.4.2’
#> 
#> Attaching package: ‘plotly’
#> The following objects are masked from ‘package:flextable’:
#> 
#>     highlight, style
#> The following object is masked from ‘package:MASS’:
#> 
#>     select
#> The following objects are masked from ‘package:plyr’:
#> 
#>     arrange, mutate, rename, summarise
#> The following object is masked from ‘package:ggplot2’:
#> 
#>     last_plot
#> The following object is masked from ‘package:stats’:
#> 
#>     filter
#> The following object is masked from ‘package:graphics’:
#> 
#>     layout
#> 
#> Attaching package: ‘viridis’
#> The following object is masked from ‘package:scales’:
#> 
#>     viridis_pal
#> Registered S3 method overwritten by 'dendextend':
#>   method     from 
#>   rev.hclust vegan
#> Registered S3 method overwritten by 'seriation':
#>   method         from 
#>   reorder.hclust vegan
#> 
#> ======================
#> Welcome to heatmaply version 1.4.2
#> 
#> Type citation('heatmaply') for how to cite the package.
#> Type ?heatmaply for the main documentation.
#> 
#> The github page is: https://github.com/talgalili/heatmaply/
#> Please submit your suggestions and bug-reports at: https://github.com/talgalili/heatmaply/issues
#> You may ask questions at stackoverflow, use the r and heatmaply tags: 
#> 	 https://stackoverflow.com/questions/tagged/heatmaply
#> ======================
#> [1] "permute"
#> [1] ‘0.9.7’
#> !!!>>>>>>>>>> your output will be in this directory: /home/hp/Documents/GitHub/GEENOME/docs/reference 
#> !!!>>>>>>>>>> make sure that this script and your data in the same directory 
#> !!!>>>>>>>>>> The output plots in details will be saved in separated folder called *GEENOME_plots*Done >>>>>>>>>>>>>>>>>>>>>>> 1.Done Preprocessing 
#> Done >>>>>>>>>>>>>>>>>>>>>>> 2.Done Explore the data 
#> `set.seed(1)` was used to initialize repeatable random subsampling.
#> Please record this for your records so others can reproduce.
#> Try `set.seed(1); .Random.seed` for the full vector
#> ...
#> 1 samples removedbecause they contained fewer reads than `sample.size`.
#> Up to first five removed samples are: 
#> Sample-56
#> ...
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Done >>>>>>>>>>>>>>>>>>>>>>> 3.Done Alpha diversity 
#> Warning: The melt generic in data.table has been passed a list and will attempt to redirect to the relevant reshape2 method; please note that reshape2 is deprecated, and this redirection is now deprecated as well. To continue using melt methods from reshape2 while both libraries are attached, e.g. melt.list, you can prepend the namespace like reshape2::melt(sub_dist). In the next version, this warning will become an error.
#> Warning: The melt generic in data.table has been passed a matrix and will attempt to redirect to the relevant reshape2 method; please note that reshape2 is deprecated, and this redirection is now deprecated as well. To continue using melt methods from reshape2 while both libraries are attached, e.g. melt.list, you can prepend the namespace like reshape2::melt(physeq_Distance). In the next version, this warning will become an error.
#> Done >>>>>>>>>>>>>>>>>>>>>>> 4.Done Beta diversity 
#> Square root transformation
#> Wisconsin double standardization
#> Run 0 stress 0.2133623 
#> Run 1 stress 0.2178022 
#> Run 2 stress 0.2211338 
#> Run 3 stress 0.2209762 
#> Run 4 stress 0.416746 
#> Run 5 stress 0.2136226 
#> ... Procrustes: rmse 0.01152948  max resid 0.1018301 
#> Run 6 stress 0.2245528 
#> Run 7 stress 0.2149561 
#> Run 8 stress 0.2136543 
#> ... Procrustes: rmse 0.01401871  max resid 0.1573673 
#> Run 9 stress 0.2175167 
#> Run 10 stress 0.2255632 
#> Run 11 stress 0.2187455 
#> Run 12 stress 0.2193386 
#> Run 13 stress 0.2274358 
#> Run 14 stress 0.2163268 
#> Run 15 stress 0.2264926 
#> Run 16 stress 0.2261255 
#> Run 17 stress 0.2267637 
#> Run 18 stress 0.2204196 
#> Run 19 stress 0.2203056 
#> Run 20 stress 0.2196194 
#> *** Best solution was not repeated -- monoMDS stopping criteria:
#>     18: stress ratio > sratmax
#>      2: scale factor of the gradient < sfgrmin
#> Square root transformation
#> Wisconsin double standardization
#> Run 0 stress 0.2133623 
#> Run 1 stress 0.2219124 
#> Run 2 stress 0.213183 
#> ... New best solution
#> ... Procrustes: rmse 0.01938682  max resid 0.1554535 
#> Run 3 stress 0.2294731 
#> Run 4 stress 0.2271144 
#> Run 5 stress 0.2182722 
#> Run 6 stress 0.2233627 
#> Run 7 stress 0.2133671 
#> ... Procrustes: rmse 0.01646453  max resid 0.1245292 
#> Run 8 stress 0.2127607 
#> ... New best solution
#> ... Procrustes: rmse 0.0174943  max resid 0.1676176 
#> Run 9 stress 0.221265 
#> Run 10 stress 0.2212965 
#> Run 11 stress 0.2183751 
#> Run 12 stress 0.2222367 
#> Run 13 stress 0.2250929 
#> Run 14 stress 0.2266284 
#> Run 15 stress 0.2202835 
#> Run 16 stress 0.2229953 
#> Run 17 stress 0.2237853 
#> Run 18 stress 0.2163237 
#> Run 19 stress 0.2174355 
#> Run 20 stress 0.2211585 
#> *** Best solution was not repeated -- monoMDS stopping criteria:
#>      2: no. of iterations >= maxit
#>     16: stress ratio > sratmax
#>      2: scale factor of the gradient < sfgrmin
#> No eigenvalues found in ordination
#> NULL
#> Warning: The following aesthetics were dropped during statistical transformation: shape
#> ℹ This can happen when ggplot fails to infer the correct grouping structure in
#>   the data.
#> ℹ Did you forget to specify a `group` aesthetic or to convert a numerical
#>   variable into a factor?
#> Warning: The following aesthetics were dropped during statistical transformation: shape
#> ℹ This can happen when ggplot fails to infer the correct grouping structure in
#>   the data.
#> ℹ Did you forget to specify a `group` aesthetic or to convert a numerical
#>   variable into a factor?
#> Warning: The following aesthetics were dropped during statistical transformation: shape
#> ℹ This can happen when ggplot fails to infer the correct grouping structure in
#>   the data.
#> ℹ Did you forget to specify a `group` aesthetic or to convert a numerical
#>   variable into a factor?
#> Square root transformation
#> Wisconsin double standardization
#> Run 0 stress 0.2133623 
#> Run 1 stress 0.2261313 
#> Run 2 stress 0.2188186 
#> Run 3 stress 0.216867 
#> Run 4 stress 0.2217179 
#> Run 5 stress 0.2159772 
#> Run 6 stress 0.2264515 
#> Run 7 stress 0.2184361 
#> Run 8 stress 0.2227609 
#> Run 9 stress 0.2298418 
#> Run 10 stress 0.2135553 
#> ... Procrustes: rmse 0.01747937  max resid 0.1571733 
#> Run 11 stress 0.225492 
#> Run 12 stress 0.2189902 
#> Run 13 stress 0.2164226 
#> Run 14 stress 0.2157921 
#> Run 15 stress 0.2233034 
#> Run 16 stress 0.2218725 
#> Run 17 stress 0.2201873 
#> Run 18 stress 0.2204809 
#> Run 19 stress 0.2255861 
#> Run 20 stress 0.2151504 
#> *** Best solution was not repeated -- monoMDS stopping criteria:
#>      2: no. of iterations >= maxit
#>     15: stress ratio > sratmax
#>      3: scale factor of the gradient < sfgrmin
#> Warning: The following aesthetics were dropped during statistical transformation: shape
#> ℹ This can happen when ggplot fails to infer the correct grouping structure in
#>   the data.
#> ℹ Did you forget to specify a `group` aesthetic or to convert a numerical
#>   variable into a factor?
#> Warning: The following aesthetics were dropped during statistical transformation: shape
#> ℹ This can happen when ggplot fails to infer the correct grouping structure in
#>   the data.
#> ℹ Did you forget to specify a `group` aesthetic or to convert a numerical
#>   variable into a factor?
#> Warning: The following aesthetics were dropped during statistical transformation: shape
#> ℹ This can happen when ggplot fails to infer the correct grouping structure in
#>   the data.
#> ℹ Did you forget to specify a `group` aesthetic or to convert a numerical
#>   variable into a factor?
#> Done >>>>>>>>>>>>>>>>>>>>>>> 5.Done ordination plots 
#> Error in adonis2(eval(parse(text = model_formula)), data = sampledf, strata = Permanova.strata,     permutations = permanova.permutation_number): object 'Permanova.strata' not found